task 7

+++ routes/web.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\TaskController;
 use App\Http\Controllers\UserController;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -22,11 +23,11 @@ Route::get('/', function () {
 // Resource routes of tasks
 Route::get('/tasks', [TaskController::class, 'index'])->name('tasks.index');
 Route::get('/tasks/create', [TaskController::class, 'create'])->name('tasks.create');
-Route::post('/tasks', [TaskController::class, 'store'])->name('tasks.store');
+Route::post('/tasks', [TaskController::class, 'store'])->name('tasks.store')->middleware('admin');
 Route::get('/tasks/{task}', [TaskController::class, 'show'])->name('tasks.show');
 Route::get('/tasks/{task}/edit', [TaskController::class, 'edit'])->name('tasks.edit');
-Route::put('/tasks/{task}', [TaskController::class, 'update'])->name('tasks.update');
-Route::delete('/tasks/{task}', [TaskController::class, 'delete'])->name('tasks.delete');
+Route::put('/tasks/{task}', [TaskController::class, 'update'])->name('tasks.update')->middleware('admin');
+Route::delete('/tasks/{task}', [TaskController::class, 'delete'])->name('tasks.delete')->middleware('admin');
 
 // Resource routes of users
 Route::resource('users', UserController::class);
